primary nav: put styles in separate file
